Complete Reference

Glimmer

Glimmer.examplesFolder — Method

examplesFolder()

return the examples folder (~/examples) for this package.

Glimmer.examplesList — Method

examplesList()

Display a list of avalable examples.

Glimmer.rootFolder — Method

rootFolder()

return the root folder for this package.

Glimmer.runExample — Method

runExample(example::String)

Run the specific example. [example] is given without extension. For example, if we have an example file named "Controls.jl" the command to run it is: Glimmer.runExample("Controls")

UIControls

Glimmer.UIControls.Slider — Type

Slider Control

Create a slider control.

    text::String = ""                   # text to apear to the left of the slider
    trailing_text::String = ""          # text to apear to the right of the slider
    min::Float64 = 0.0                  # minimum value
    max::Float64 = 100.0                # maximum value
    value::Float64 = 0                  # starting value (unused if connected to a variable)
    step::Float64 = 1.0                 # tick change value
    variable::Any = nothing             # variable

FlexUI

Glimmer.FlexUI.forceUpdateControls! — Method

forceUpdateControls!(app::App)

Force an update of controls that are not updated during the normal execution of an application. Currently, RawHTML is the only component that behaves like that due to the time it can take to update it.