Complete Reference

Glimmer

Glimmer.examplesFolder — Method

examplesFolder()

return the examples folder (~/examples) for this package.

Glimmer.examplesList — Method

examplesList()

Display a list of avalable examples.

Glimmer.rootFolder — Method

rootFolder()

return the root folder for this package.

Glimmer.runExample — Method

runExample(example::String)

Run the specific example. [example] is given without extension. For example, if we have an example file named "Controls.jl" the command to run it is: Glimmer.runExample("Controls")

UIControls (Glimmer.UIControls])

Glimmer.UIControls.Slider — Type

Slider Control

Create a slider control.

    text::String = ""                   # text to apear to the left of the slider
    trailingText::String = ""           # text to apear to the right of the slider
    min::Float64 = 0.0                  # minimum value
    max::Float64 = 100.0                # maximum value
    value::Float64 = 0                  # starting value (unused if connected to a variable)
    step::Float64 = 1.0                 # tick change value
    variable::Any = nothing             # variable

FlexUI (Glimmer.FlexUI)

Glimmer.FlexUI.forceUpdateControls! — Method

forceUpdateControls!(app::App)

Force an update of controls that are not updated during the normal execution of an application. Currently, RawHTML is the only component that behaves like that due to the time it can take to update it.

Glimmer.FlexUI.variable — Method

variable(app::App, name::String)

Return the variable object or nothing if variable cannot be found in app.

UI Variables (Glimmer.UIVariables)

Glimmer.UIVariables.addGridColumn! — Method

addGridColumn!(var::Variable, colDef::Dict)

Add a new grid column definition as the last column.

Glimmer.UIVariables.clearGridColumns! — Method

clearGridColumns!(var::Variable)

Clear all existing definitions of columns for this var grid.

Glimmer.UIVariables.deleteGridColumn! — Method

deleteGridColumn!(var::Variable, index::Int64)

Delete a grid column at a specific index.

Glimmer.UIVariables.deleteGridColumn! — Method

deleteGridColumn!(var::Variable, name::String)

Delete a grid column at a specific index.

Glimmer.UIVariables.emit — Function

on(func::Function, var::AbstractUIVariable, event::Symbol = :valueChanged)

Emits the specific event on a variable, calling all the listeners.

Glimmer.UIVariables.forEachGridColumn! — Method

forEachGridColumn!(var::Variable, option::String, value)

Set a column options for each exiting column definition in this grid variable.

Glimmer.UIVariables.gridColumnOption! — Method

gridColumnOption!(var::Variable, colName::String, option::String, value)

Set a grid column's option.

Glimmer.UIVariables.gridDefaultColDef! — Method

gridDefaultColDef!(var::Variable, colDef::Dict)

Set the default column definition for this grid variable.

Glimmer.UIVariables.gridOption! — Method

gridOption!(var::Variable, option::String, value)

Set a grid options.

Glimmer.UIVariables.insertGridColumn! — Method

insertGridColumn!(var::Variable, index::Int64, colDef::Dict)

Add a new grid column definition as the last column.

Glimmer.UIVariables.on — Function

on(func::Function, var::AbstractUIVariable, event::Symbol = :valueChanged)

Adds function func as listener to the variable. Whenever variables's value is set via var[] = val or one of the defined UI Controls, func is called with val. Most variables have just one type of events associated with them: valueChanged. But some, such as 'aggrid' have additional events that you can bind listeners to.

Glimmer.UIVariables.rawValue — Method

rawValue(var::Variable) -> string

returns the internal value of the variable. in most cases this value would be identical to using var[], but in some cases, such as aggrid type variable, this would return the internal structure of the grid definition, allowing the user to manually manipulate it for better control on the grid behevior.

Glimmer.UIVariables.replaceGridColumn! — Method

replaceGridColumn!(var::Variable, index::Int64, colDef::Dict)

Replace an existing grid column definition with a new one.

Grid Utilities (Glimmer.GridUtils)

Glimmer.GridUtils.getTypes — Method

getTypes(m::Any, arr::Vector{Any})

This function has no place in this package and is just a utility function that fills an array with information of types in the given module. The function is here in order to test the agGrid tree view capabilities but it has nothing to do with UI or Visualization.

Glimmer.GridUtils.table2agGrid — Method

table2agGrid(table)

Takes a Tables.jl compatable table and return the agGrid represantation of it with default colDefs.

Glimmer.GridUtils.table2agGridTree — Method

table2agGridTree(table, id::String, parentId::String)

Takes a Tables.jl compatable table and return the agGrid represantation of it with default colDefs. The table data is assumed to represent an hericical data structure where the id parameter is the field containing a unique id, and the parentId parameter is the field containing the parent id of the specific row.

Glimmer.GridUtils.toJSONTable — Function

toJSONTable(data::Vector{Any}, parent_id::Int64 = -1)

Utility function to convert the result of getTypes in to a JSON table (Tables.jl compatiable data structure)

INTERNAL: Julia to JavaScript Bridge (Glimmer.JuliaJSBridge)

Glimmer.JuliaJSBridge.JS_loader_node — Method

JS_loader_node(filenames...)

Return a WebIO.Node object that contains all the supplied files as children nodes. Recognized extentions files such as '.js' and '.css' are contains in a :script and :style nodes respectivly.

Glimmer.JuliaJSBridge.fixMaterialIConsStyles — Method

Fixing the path in the styles file for the MaterialIcons font for Blink to load locally from the dist directory